split steam_user_stats class into multiple files for readability
